fix bug: shop-reducer overrides localStorage in CartLayout

+++ fruits_in_choco_frontend/src/components/Admin/ProductPage/AdminProductPage.js
@@ -98,7 +98,9 @@ const AdminProductPage = props => {
     }
 
     const remove = (productId, ratingId) => {
-        props.deleteReview(productId, ratingId);
+        if (window.confirm('Удлаить?')) {
+            props.deleteReview(productId, ratingId);
+        }
     }
 
     return <>
